+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * #214 Cleaning up APIs deprecated since <= 0.25
 * #215 Remove support for date/time values as arrays of components
 * #216 Config: addressing array elements via property paths
+* #218 Percent value object
 
 ## 0.25